model_api/ImageModel/requirements.txt
fastapi
uvicorn
requests
pillow
pydantic